fix: wrong type of type when keyboard hides
